Configure MySQL database and display /users.
Note: Remember to run `php artisan` commands from within an SSH session with the
Homestead vm as opposed to the local terminal session. This is because artisan
attempts to access the MySQL server using the configuration info found in
app/config/database.php, and the port number used to access the MySQL server is
different depending on whether you log in from the host computer or the guest vm.